mail-processor/requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.4
openpyxl==3.1.2 